REORG/MAJOR: Separated match classes and logic logically and amended the cache implementation
These changes improve code maintenance and readability without altering logic. Where some functionality was not implemented compared to C# this has been added.

1. Match

Methods related to the matching process were inconsistently placed in the Controller class and the Match class. This refactoring revises the MatchState class to include all working fields used during the process of device detection. Any methods related specifically to MatchState members and the matching process are now placed in this revised class. Other methods related to matching have been moved to static implementations in the Controller. The Controller class now only contains static members.

The Match class is now purely a public class for the purposes of retrieving the results of device detection. It can be populated from;

a) a fresh detection calculation where the MatchState contains the results; or
b) the results of a previous detection held in the cache.

The shallow copying of values previously required to retrieve results from the cache has been replaced with a single MatchResult class. This is an immutable class used to place match results into the cache and be shared in by multiple threads. Both MatchState and MatchResult implemented the common interface IMatchResult. As such they can be used interchangeably for the purposes of returning results. Only MatchResult can be placed in the cache.

Handling of multiple HTTP headers has changed to make use of the MatchState class. This change improved readability.

2. Cache

A single set of Cache classes in the cache package has been created to replace the previous ones. Generic types are used to avoid needing to create multiple implementations.
The cache model can now easily be altered to other methods if required in the future.

The ICacheSource interface is used to pass an underlying source of cache items when a request is made. Should the key not be found this source can be used to retrieve the value. MatchState and the stream lists implement this interface.

The User-Agent cache and the stream lists classes which all make use of the cache class have been altered to make use of the implementation.

3. Other Changes

The elapsed time in nanoseconds for the match is now implemented. A provider can be created which will set this value.

The NodeStreamFactory and ProfileStreamFactory no longer require a pool instance in their constructor as the pool is no longer needed by their implementation.

Unused methods and fields have been removed.

.gitignore has changed to remove some unneeded NetBeans working files.

Minor documentation changes include setting the copyright year to 2015.