REFACT/MAJOR: Altered the Cache, Match and Profile classes plus others
IMatchResult removed and replaced by using MatchResult as a base class for MatchState.

Throws IOException added to all base class methods to avoid needing to bury IOExceptions. Three places remain where IOException can’t be thrown and the new WrappedIOException needs to be added.

Profile class no longer maintains a separate index for property names. This duplicated functionality with the PropertiesList class and has been removed to save memory and initialisation time.

The Profile class’ index of property indexes to Values is now initialised when the MemoryFactory is instructed to initialise all data. This ensures more memory is fully allocated when fully initialising the DataSet.

Cache can now be constructed with a default loader and the get method used without providing a loader. This is more efficient when used to cache data from the Stream lists.

Commented out methods and unused methods have been removed.