REORG/MAJOR: Modified AutoUpdate to simpler implementation
Changed the Pool and Source so that the Pool is responsible for closing the readers and not the source.

Implemented changes for MemoryMapped files to stand a better chance freeing the lock faster. NOTE: memory mapped files will be replaced in a future version of this branch.

Refactored AutoUpdate to provide a simpler implementation.

Added basic tests for AutoUpdate.

NOTE; These changes are not intended to be applied to develop at this stage.