REORG/MINOR: Moved source classes and fixed AutoUpdate
The source classes have been moved to the readers package as they are closely associated with BinaryReader and used by more than just the stream package. The SourceFile class no longer maintains a list of file handles and instead creates a fresh memory mapped byte buffer only when a new stream is required.

Pool no longer is closable as it is the callers responsibility to free the source.

The AutoUpdate tests have been fixed to ensure files are deleted correctly.