OPTIM/MINOR: LRU cache implementation
The revised cache uses an LRU design to avoid problems with servicing the cache and inconsistent behaviour when operated in a multi threaded environment. The interface has also been simplified so that the cache loader is no longer a 3rd type.

Added suppression for volatile array and double checked locked warnings.

Note: Test class needs to be added.