OPTIM/MINOR: Binary search consolidation
All the binary search operations are now consolidated into the search package providing a single implementation of the algorithm for use in all places where the standard Java binarySearch method can not be used because the items in the list have different key types to the item.

An initial integration test has been added to validate the dataSet findProfile method which is now modified to use these core methods.