BUG/MAJOR: Fixed errors in Numeric and Closest matching logic
The following two changes address problems with Numeric and Closest matching logic.

Node class now includes values between 1,000 and 10,000 in numeric calculation. The Range class now works identically to C# and changes the upper value is lower than only.

Changed ClosestScore logic to use BaseEntity static methods for numeric calculations. This also addressed problems with the logic in the duplicated functions in ClosestScore which were returning incorrect results.

The following changes relate to match metrics or code quality.

BinarySearch now returns the number of iterations used as well as the index of the result. This is needed to populate matching metrics correctly. The performance impact of the change needs to be gauged and possibly altered to support two BinarySearch implementations one to count the iterations and the other to just return the index.

MostFrequentFilter now supports the max result parameter which limits the number of results. This has the added benefit of reducing processing time as items are not needlessly added to the list.

RankedSignatureIterator now returns the number of items in the iteration to aid in calculating matching metrics.

Match metrics now support NodesFound to return the number of nodes matched to aid diagnosing problems.

Minor code layout issues addressed including the placing of && and || on the preceeding line of the IF statement.