BUG: V31 signature now creates it a fresh ArrayList for nodeOffsets.
The V31 signature was previously using a copy of the shared ArrayList provided by the BinaryReader. This resulted in each Signature using a reference to the same list of NodeOffsets which in turn resulted in a failure in the basic device detection logic. This implementation no longer uses the shared ArrayList in BinaryReader and instead creates its own list. The method for returning lists of non negative integers has been modified to provide better comments and parameter names. Unused parameters have also been removed.

A defect in the Node base class which did not synchronise fetching the parent node has been addressed.