FEAT: Added support for data files which include JavaScript which overrides property value results. For example; returning the screen width in pixels based on a JavaScript snippet rather than using the default value. The commit also deprecates fully the getResults method of Provider and related calling classes.
From July 2016 all 51Degrees data files will contain additional JavaScript properties which can be added to client code to set cookies that will override static values. For example; if a desktop computer can have multiple screen sizes then the actual screen size can be calculated dynamically for each session and overrides the static value. This concept is termed Property Value Override and is implemented via Dynamic values alongside existing Static values.

By convention Properties that implement this feature have a second property which shares the same prefix and is appended with the suffix JavaScript. The property has the type JavaScript. For example; the property ScreenPixelsWidth now relates to ScreenPixelsWidthJavaScript.

These JavaScript properties will be assigned to the Category “Property Value Override”. Other JavaScript properties that aren’t assigned to the Category are ignored by the feature.

If after initial device detection a value is available for this new JavaScript property, the value can be inserted into the core.js file. The JavaScript will then execute and set a cookie with the name 51D_XXXX where XXXX is the name of the property. For example; cookie 51D_ScreenPixelsWidth will relate to the ScreenPixelsWidth property.

Where such a cookie is present in the HTTP request the server will favour returning the cookie value rather than the static value from the data set. In this way any property that may not be know for certain devices and browsers when the data file is generated by 51Degrees can be set dynamically based on additional JavaScript.

The Value class can now be constructed from a string and the Values class has a public constructor to support dynamic properties.

Summary of Class Changes:

Constants – added constants for cookie handling:

 PROPERTY_VALUE_OVERRIDE_COOKIE_PREFIX – the prefix to apply to the cookie name. Must match that used by 51Degrees in the data file.

 PROPERTY_VALUE_OVERRIDE_CATEGORY – the category that JavaScript properties must be assigned to in the data file to be included in the Property Value Override feature.

DataSet – now includes an internal property called getJavaScriptProperties and getPropertyValueOverrideProperties which returns all the properties that have the type JavaScript. Used to quickly narrow the list of possible items when filtering by JavaScript type.

Value & Values – these classes can now support dynamic as well as static values.

Match – when a collection of HTTP headers is provided as evidence for the match the “cookie” HTTP header is retained for future analysis. When a property value is requested the cookie value is analysed for the presence of any values. If a value is available (an override) then it will be returned in place of the value that is available in the static data set.

In order to support the new functionality, the Reset method is always called to ensure that the results of a previous match instances cookies collection can be not returned by mistake. This also addresses an unreported bug with the Profile Overrides which could suffer from the same issue in a custom 3rd party implementation.

Provider – sets the cookie after the 51Degrees device detection has completed using either a single HTTP header, or multiple HTTP headers.

JavaScriptProvider – now modified to include the Property Value Override JavaScript when available.

BaseServlet – modified to no longer use deprecated method getResults.

Bandwidth, ImageOptimiser, ProfileOverride - These feature classes in WebApp have been changed to no longer use the deprecated method getResults.

Example – the example web page has been modified to include the 51D/core.js to demonstrate ScreenPixelsWidth.

New Classes:

PropertyValueOverrides – contains the functionality to add the JavaScript snippets to the core.js file.

This commit closes the following work items:

Changed version to 3.2.5-SNAPSHOT.

Tests:

TODO - Need to be added to support the changes to the Match class.

Closes the following tasks:

Closes #753
Closes #755

The commit partially relates to the progression of:

#757 - tests will be needed following peer review of the implementation.