FEAT: Usage sharing now operates when enabled only when a new session is being created. Enhancements have been made to the background thread processing.
Request information is only considered for sharing during the first request which is indicated by the isNew property of the session associated with the request.

The background thread and controller are now associated directly with the WebProvider instance and closed when the provider is closed.

Simplified the process of compiling the XML fragments.

Write data in UTF-8 encoding.

Modified methods that don’t need to be public so that they are package local.