BUG/TEST/MINOR: Respond to 304 responses where the data file does not need to be updated. Auto update responds to 301 and 302 redirections.
The AutoUpdate class now correctly sets the If-Modified-Since header and interprets the Last-Modified response header to set the date time of the file correctly. This change prevents the data file from being download multiple times when it has not changed compared to the prior version.

The tests have been modified to check for the correct last modified response.

Added logic to check for redirect responses to future proof should 51Degrees change the URL of the auto update service.

Tidied the WebProvider to ensure that log message appear correctly and code formatting is improved.

Auto update logger messages are more comprehensive in the webapp.