EXAMPLE: Added all profiles example.
The all profiles example fetches all the hardware profiles, then outputs all the devices and their properties to a CSV file. Two filters are applied to the properties that are printed: firstly only hardware properties are fetched as they are the only properties that relate to a hardware profiles, also any javascript property is given the value "N/A" as these can return long multiline javascript which will ruin any CSV file formatting.