BUG: Fix for profile override cookie.
The pipe character being used to split multiple profile ids was being treated as a regex "or" meanging every digit of the id was split. The pipe character is now properly escaped so that it is treated as a character.

In WebProvider, the profile override logic present in the depricated getResult method has been transefered to the match method to enable the feature.