FEAT/BUG: isTempFile flag in DataSet constructor is now used, and a default constructor provided.
The flag was previously not used, false was always given to the SourceFile constructor. The Flag is now passed to the SourceFile constructor. In addition, a constructor which takes only the file path as an argument has been added, this calls the same function with the isTempFile flag set to false